requirements.txt
# Discord library with voice support
nextcord[voice]

# Audio processing
pynacl
cffi

# Python utilities
setuptools
wheel
gtts
python-dotenv